Move 'src/domain/cards/*' to 'src/common/*'
