+++ src/common/applicationDetails/ApplicationDetails.test.jsx
