+++ src/common/applicationDetails/applicationDetails.module.scss
