+++ src/common/berthDetails/berthDetails.module.scss
