+++ src/domain/harbors/harborDetails/HarborDetails.tsx
@@ -8,7 +8,7 @@ import Section from '../../../common/section/Section';
 import Text from '../../../common/text/Text';
 import { formatDimension } from '../../../common/utils/format';
 import { HarborData } from '../types';
-import HarborMapLinks from '../../cards/harborMapLinks/HarborMapLinks';
+import HarborMapLinks from '../../../common/harborMapLinks/HarborMapLinks';
 
 export type HarborDetailsProps = {
   imageFile: HarborData['imageFile'];
