+++ src/domain/individualApplication/IndividualApplicationPage.tsx
@@ -6,7 +6,7 @@ import { Notification } from 'hds-react';
 import styles from './individualApplicationPage.module.scss';
 import Card from '../../common/card/Card';
 import CardBody from '../../common/cardBody/CardBody';
-import ApplicationDetails, { ApplicationDetailsProps } from '../cards/applicationDetails/ApplicationDetails';
+import ApplicationDetails, { ApplicationDetailsProps } from '../../common/applicationDetails/ApplicationDetails';
 import CardHeader from '../../common/cardHeader/CardHeader';
 import Table, { Column } from '../../common/table/Table';
 import CustomersTableTools, { CustomersTableToolsProps } from './customersTableTools/CustomersTableTools';
@@ -14,7 +14,7 @@ import Text from '../../common/text/Text';
 import { formatDate } from '../../common/utils/format';
 import Chip from '../../common/chip/Chip';
 import { APPLICATION_STATUS } from '../../common/utils/consonants';
-import CustomerProfileCard, { CustomerProfileCardProps } from '../cards/customerProfileCard/CustomerProfileCard';
+import CustomerProfileCard, { CustomerProfileCardProps } from '../../common/customerProfileCard/CustomerProfileCard';
 import OfferCard, { OfferCardProps } from './offerCard/OfferCard';
 import { OrganizationType } from '../../@types/__generated__/globalTypes';
 import Pagination, { PaginationProps } from '../../common/pagination/Pagination';
